requirements.txt
selenium==4.7.2
pytest==7.2.0
allure-pytest==2.12.0

